# docs/api.yaml
openapi: 3.1.0
info:
  title: HyperHarbor Host API
  version: 1.1.0
  summary: Manage Hyper-V virtual machines on a HyperHarbor host.
  description: |
    Contract between the HyperHarbor host service and HyperHarbor clients.

    Transport and security:
    - All endpoints are served over HTTPS using the host's self-signed certificate.
    - Every operation requires mutual TLS with a client certificate issued during
      pairing, except the pairing request and confirm operations, which are
      authenticated by the PIN-based key exchange described under the Pairing tag.
    - JSON property names are camelCase. Enumerations are serialized as strings.
    - Errors use RFC 7807 problem details (`application/problem+json`).
    - Every operation that changes state is recorded in the host's audit log with the User,
      device, virtual machine, and outcome. If the entry cannot be written, the operation is
      refused with 503.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: https://{host}:{port}/api/v1
    description: HyperHarbor host
    variables:
      host:
        default: hyperharbor.local
        description: Host name or IP address discovered via mDNS or entered manually.
      port:
        default: "48443"
        description: API port advertised in the mDNS TXT record.
security:
  - mutualTLS: []
...
paths:
...
  /vms:
    get:
      tags: [VMs]
      operationId: listVms
      summary: List virtual machines on the host.
      responses:
        "200":
          description: All virtual machines visible to the host service.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vm"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/HyperVUnavailable"
  /vms/{vmId}:
    parameters:
      - $ref: "#/components/parameters/VmId"
    get:
      tags: [VMs]
      operationId: getVm
      summary: Get a single virtual machine.
      responses:
        "200":
          description: The virtual machine.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vm"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/HyperVUnavailable"
  /vms/{vmId}/actions:
    parameters:
      - $ref: "#/components/parameters/VmId"
    post:
      tags: [VMs]
      operationId: performVmAction
      summary: Request a power action on a virtual machine.
      description: |
        Starts an asynchronous state change. Poll `GET /vms/{vmId}` (the `Location`
        header) to observe the resulting state.

        `turnOff` is equivalent to pulling the power cord and can lose unsaved data
        in the guest. Clients must confirm with the user before sending it.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VmActionRequest"
            example:
              action: start
      responses:
        "202":
          description: The action was accepted.
          headers:
            Location:
              description: URL of the virtual machine resource.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VmActionResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "502":
          $ref: "#/components/responses/HyperVOperationFailed"
        "503":
          $ref: "#/components/responses/HyperVUnavailable"
  /vms/{vmId}/provision:
    parameters:
      - $ref: "#/components/parameters/VmId"
    post:
      tags: [VMs]
      operationId: provisionVm
      summary: Create the calling User's local account in a VM.
      description: |
        One-time setup for one-click Remote Desktop. Creates (or reuses) the User's local
        account, for example `hh-owner`, with the given guest administrator credential, and
        optionally enables Remote Desktop. The VM is marked provisioned only after the account
        is verified.

        - Windows (10 / Server 2016 or later): PowerShell Direct; the account is added to
          Remote Desktop Users.
        - Linux: SSH with password authentication to the VM's address, then sudo; Remote
          Desktop is xrdp. The guest's SSH host key is pinned on first use and checked on
          every later connection, including when the VM is set up again (unless
          `trustNewHostKey` is true). Installing packages can take several minutes.

        The guest OS comes from `Vm.guestOs`; a VM whose family is `unknown` cannot be provisioned.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProvisionVmRequest"
      responses:
        "200":
          description: The account exists and was verified.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VmProvisioning"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: |
            The VM is not running, its guest OS is unknown or has no address (Linux), an
            existing account with the same name is not a local account, or (Linux) the guest's
            SSH host key differs from the pinned one and `trustNewHostKey` is false.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "422":
          description: The guest rejected the administrator credential.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "502":
          $ref: "#/components/responses/GuestOperationFailed"
        "503":
          $ref: "#/components/responses/GuestUnavailable"
  /vms/{vmId}/connect:
    parameters:
      - $ref: "#/components/parameters/VmId"
    post:
      tags: [VMs]
      operationId: connectVm
      summary: Get Remote Desktop credentials for the calling User's account.
      description: |
        Rotates the User's account password in the guest through PowerShell Direct and
        returns it. Rotations are serialized per VM and User; requests within the reuse
        window (60 seconds by default) receive the same password. The response is sent
        with `Cache-Control: no-store`.
      responses:
        "200":
          description: Credentials for Remote Desktop.
          headers:
            Cache-Control:
              schema:
                type: string
                const: no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VmConnection"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: The VM is not provisioned for this User, is not running, or has no address.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "502":
          $ref: "#/components/responses/GuestOperationFailed"
        "503":
          $ref: "#/components/responses/GuestUnavailable"
...
  /wake/info:
    get:
      tags: [Wake]
      operationId: getWakeInfo
      summary: Get network adapter details needed to wake the host.
      responses:
        "200":
          description: Wake-capable adapters.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WakeInfo"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /wake/readiness:
    get:
      tags: [Wake]
      operationId: getWakeReadiness
      summary: Check whether the host is configured to wake on LAN.
      responses:
        "200":
          description: Readiness check results.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WakeReadiness"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /wake/readiness/fix:
    post:
      tags: [Wake]
      operationId: fixWakeReadiness
      summary: Apply automatic fixes for failed readiness checks.
      description: |
        Fixes change system settings and need administrator rights. When the host service
        has them, the fixes are applied immediately (200). Otherwise the host tray asks the
        user to approve them with a UAC prompt (202); poll `GET /wake/readiness` to see the
        result. Only checks with `autoFixAvailable` can be fixed.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WakeFixRequest"
      responses:
        "200":
          description: Readiness check results after applying fixes.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WakeReadiness"
        "202":
          description: The user at the host has been asked to approve the fixes.
          headers:
            Location:
              description: URL of the readiness resource to poll.
              schema:
                type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          description: Approval is needed and no tray app is running on the host to ask.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /wake/test:
    post:
      tags: [Wake]
      operationId: startWakeTest
      summary: Put the host to sleep after a delay so the client can test waking it.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WakeTestRequest"
      responses:
        "202":
          description: The host will sleep at the indicated time.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WakeTestScheduled"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
components:
  securitySchemes:
    mutualTLS:
      type: mutualTLS
      description: Client certificate issued to the device during pairing.
  parameters:
    VmId:
      name: vmId
      in: path
      required: true
      description: Hyper-V virtual machine ID.
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: The request was malformed or failed validation.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    Unauthorized:
      description: The client certificate is missing or not paired with this host.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    NotFound:
      description: The resource does not exist.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    Conflict:
      description: The operation is not valid in the current state.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    HyperVOperationFailed:
      description: Hyper-V rejected the request with an unexpected return code.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    GuestOperationFailed:
      description: An operation in the guest (PowerShell Direct or SSH) failed.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    GuestUnavailable:
      description: |
        The guest is not ready: PowerShell Direct cannot reach it (still starting, or integration
        services are off), or SSH on port 22 does not answer (Linux).
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    HyperVUnavailable:
      description: Hyper-V is not enabled or the host service cannot access it.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
...
  schemas:
    ProblemDetails:
      type: object
      description: RFC 7807 problem details.
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
...
    VmState:
      type: string
      enum: [running, off, saved, paused, starting, stopping, saving, pausing, resuming, other]
    GuestOsFamily:
      type: string
      description: |
        Operating system family the guest reports through Hyper-V data exchange (KVP).
        `unknown` when the guest is not running, has not finished starting, or does not run
        the data exchange integration service.
      enum: [unknown, windows, linux]
    VmGuestOs:
      type: object
      required: [family, name]
      properties:
        family:
          $ref: "#/components/schemas/GuestOsFamily"
        name:
          type: [string, "null"]
          description: Name and version the guest reports, for example "Windows 11 Pro" or "Ubuntu 24.04".
    VmAction:
      type: string
      description: |
        - `start`: power on or resume from saved state.
        - `shutdown`: graceful guest shutdown via integration services.
        - `turnOff`: immediate power off. Can lose unsaved guest data.
        - `save`: save memory state to disk and stop.
        - `restart`: graceful guest restart.
      enum: [start, shutdown, turnOff, save, restart]
    Vm:
      type: object
      required: [id, name, state, generation, rdpAvailable, ipAddresses, provisioned, remoteDesktop, guestOs]
      properties:
        id:
          type: string
          format: uuid
          description: Hyper-V virtual machine ID.
        name:
          type: string
        state:
          $ref: "#/components/schemas/VmState"
        uptimeSeconds:
          type: integer
          format: int64
          minimum: 0
        cpuUsagePercent:
          type: [integer, "null"]
          minimum: 0
          maximum: 100
        memoryAssignedMb:
          type: [integer, "null"]
          format: int64
          minimum: 0
        generation:
          type: integer
          enum: [1, 2]
        rdpAvailable:
          type: boolean
          description: |
            True when the VM accepted a TCP connection on port 3389 from the host. A VM on an
            internal or NAT switch can be available to the host but unreachable from other devices.
        ipAddresses:
          type: array
          items:
            type: string
        provisioned:
          type: boolean
          description: |
            True when the calling device's User has a verified local account on this VM, so
            `POST /vms/{vmId}/connect` can be used.
        remoteDesktop:
          $ref: "#/components/schemas/VmRemoteDesktop"
        guestOs:
          $ref: "#/components/schemas/VmGuestOs"
      example:
        id: 0b9a6f53-1c2d-4e8f-a1b2-3c4d5e6f7a8b
        name: Dev Workstation
        state: running
        uptimeSeconds: 3600
        cpuUsagePercent: 12
        memoryAssignedMb: 8192
        generation: 2
        rdpAvailable: true
        ipAddresses: [192.168.1.50]
        provisioned: true
        remoteDesktop:
          address: 192.168.1.50
          reachableFromHost: true
        guestOs:
          family: windows
          name: Windows 11 Pro
    VmRemoteDesktop:
      type: object
      required: [address, reachableFromHost]
      properties:
        address:
          type: [string, "null"]
          description: The guest address clients connect to, or null when the guest reports none.
        reachableFromHost:
          type: boolean
          description: Same as Vm.rdpAvailable.
    ProvisionVmRequest:
      type: object
      required: [adminUserName, adminPassword]
      properties:
        adminUserName:
          type: string
          minLength: 1
          description: |
            Windows: a local administrator, for example "Administrator" or ".\\admin".
            Linux: an account that can sign in over SSH with a password and use sudo.
        adminPassword:
          type: string
          minLength: 1
          description: Stored on the host with DPAPI for later password rotation. Never logged.
        enableRemoteDesktop:
          type: boolean
          default: true
          description: |
            Windows: also enable Remote Desktop and its firewall rule. Linux: also install, enable,
            and open the firewall for xrdp.
        installDesktop:
          type: boolean
          default: false
          description: |
            Linux only: install a lightweight desktop (Xfce) when the guest has none. Remote Desktop
            on Linux needs a desktop environment. Ignored for Windows guests.
        trustNewHostKey:
          type: boolean
          default: false
          description: |
            Linux only: accept an SSH host key that differs from the one pinned when the VM was
            last set up, for example after the guest was reinstalled. When false, a changed key
            fails with 409 and nothing is sent to the guest after the key exchange. Ignored for
            Windows guests.
    VmProvisioning:
      type: object
      required: [vmId, accountName, provisionedAt]
      properties:
        vmId:
          type: string
          format: uuid
        accountName:
          type: string
          description: The User's local account in the guest, for example hh-owner.
        provisionedAt:
          type: string
          format: date-time
    VmConnection:
      type: object
      required: [userName, password, address, port, expiresAt, guestOs]
      properties:
        userName:
          type: string
          description: Account to sign in with, for example ".\\hh-owner" (Windows) or "hh-owner" (Linux).
        password:
          type: string
          description: |
            Current password for the account. Valid until the next rotation; requests within
            the reuse window receive the same password. Clients must not persist it.
        address:
          type: string
        port:
          type: integer
          default: 3389
        expiresAt:
          type: string
          format: date-time
          description: End of the reuse window. The password stays valid until the next connect after it.
        guestOs:
          $ref: "#/components/schemas/GuestOsFamily"
          description: Linux guests use xrdp, which signs in over TLS instead of CredSSP (NLA).
    VmActionRequest:
      type: object
      required: [action]
      properties:
        action:
          $ref: "#/components/schemas/VmAction"
    VmActionResult:
      type: object
      required: [vmId, action, accepted, state]
      properties:
        vmId:
          type: string
          format: uuid
        action:
          $ref: "#/components/schemas/VmAction"
        accepted:
          type: boolean
        state:
          $ref: "#/components/schemas/VmState"
      example:
        vmId: 0b9a6f53-1c2d-4e8f-a1b2-3c4d5e6f7a8b
        action: turnOff
        accepted: true
        state: stopping
...
    WakeAdapter:
      type: object
      required: [name, macAddress, ipv4Address, broadcastAddress]
      properties:
        name:
          type: string
        macAddress:
          type: string
          pattern: "^([0-9A-F]{2}-){5}[0-9A-F]{2}$"
        ipv4Address:
          type: string
          format: ipv4
        broadcastAddress:
          type: string
          format: ipv4
    WakeInfo:
      type: object
      required: [adapters]
      properties:
        adapters:
          type: array
          items:
            $ref: "#/components/schemas/WakeAdapter"
      example:
        adapters:
          - name: Ethernet
            macAddress: 00-1A-2B-3C-4D-5E
            ipv4Address: 192.168.1.10
            broadcastAddress: 192.168.1.255
    WakeCheckStatus:
      type: string
      enum: [pass, warn, fail]
    WakeCheck:
      type: object
      required: [id, title, status, autoFixAvailable]
      properties:
        id:
          type: string
          description: |
            Stable check identifier: `wiredAdapter`, `nicWakeOnMagicPacket`,
            `nicAllowWake`, `sleepKeepsNetwork`, or `fastStartupDisabled`. Clients should
            show unknown identifiers using `title` and `detail`.
        title:
          type: string
        status:
          $ref: "#/components/schemas/WakeCheckStatus"
        detail:
          type: string
        autoFixAvailable:
          type: boolean
    WakeReadiness:
      type: object
      required: [ready, checks]
      properties:
        ready:
          type: boolean
          description: True when no check has status `fail`.
        checks:
          type: array
          items:
            $ref: "#/components/schemas/WakeCheck"
      example:
        ready: false
        checks:
          - id: nicWakeOnMagicPacket
            title: Network adapter wakes on magic packet
            status: fail
            detail: Wake on Magic Packet is disabled on adapter Ethernet.
            autoFixAvailable: true
    WakeFixRequest:
      type: object
      required: [checkIds]
      properties:
        checkIds:
          type: array
          minItems: 1
          items:
            type: string
    WakeTestRequest:
      type: object
      required: [delaySeconds]
      properties:
        delaySeconds:
          type: integer
          minimum: 5
          maximum: 300
    WakeTestScheduled:
      type: object
      required: [sleepAt]
      properties:
        sleepAt:
          type: string
          format: date-time